revisi :
+ kolom sisa pada pemesanan / pemesanan material
- modul status pengiriman
- fitur otomatis update status kendaraan
+ order data dari yang terbaru ke yg lama
+ perbaikan beberapa kode

+++ app/Models/Opname.php
@@ -39,7 +39,7 @@ class Opname extends Model
      */
     protected $casts = [
         'bahan_baku_id' => 'integer',
-        'volume_opname' => 'integer',
+        'volume_opname' => 'double',
         'keterangan' => 'string'
     ];
