+++ database/migrations/2018_03_12_123238_add_column_sisa_to_bahan_bakus_table.php
@@ -14,7 +14,7 @@ class AddColumnSisaToBahanBakusTable extends Migration
     public function up()
     {
         Schema::table('bahan_bakus', function (Blueprint $table) {
-            $table->integer('sisa')->nullable();
+            $table->double('sisa')->nullable();
         });
     }
