+++ database/migrations/2018_04_16_135720_create_pemesanan_bahan_bakus_table.php
@@ -18,7 +18,7 @@ class CreatePemesananBahanBakusTable extends Migration
             $table->string('nama_supplier');
             $table->integer('cp_supplier');
             $table->integer('bahan_baku_id')->unsigned();
-            $table->integer('volume_pemesanan');
+            $table->double('volume_pemesanan');
             $table->date('tanggal_pemesanan');
             $table->string('keterangan');
             $table->timestamps();
